requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.0
asyncpg>=0.29.0
aiosqlite>=0.19.0

# HTTP Client
httpx>=0.26.0

# RSS/Scraping
feedparser>=6.0.10
beautifulsoup4>=4.12.0
lxml>=5.1.0

# AI
openai>=1.10.0

# AWS
boto3>=1.34.0

# Scheduling
apscheduler>=3.10.4

# Configuration
pyyaml>=6.0.1

# Utilities
python-dotenv>=1.0.0

# Logging
python-json-logger>=2.0.7
